requirements.txt
cryptography
PySocks
